SIMPLE ACQUIRE ANALYZE VISUALIZE (AAV) PROJECT
This simple AAV project contains a LabVIEW application that acquires a sine wave signal, analyzes it based on some statistical analysis tools (Arithmetic mean) and display the output. The Express VIs for the each stage of execution are shown on the block diagram. You can run the front panel to see how the graphical data flow executes.