FIRST LabVIEW Application
This Virtual Instrument mimics an oscilloscope. Here, it accepts input as time in seconds, it separates it into hours, minutes and seconds. If the time exceeds one hour (3600 input value), then the boolean datatype LED becomes ON (mechanically acting as a switch when ON) and gives outpus as 'Delay 1 hour or longer'. However, if the value is less than the required value, it becomes off and outputs the value 'Delay less than 1 hour'.