CREATING MY FIRST APPLICATION
In this virtual instrument program, I created my first standard VI that accepts inputs of temperature measurement and displays it graphically. This application allows the user to make changes and determine the output of the program. With the adjust knob, you can delay the time multiplier as see result as a temperature graph.